+++ packages/stream_chat/lib/src/core/models/attachment_file.dart
@@ -3,7 +3,6 @@ import 'dart:typed_data';
 import 'package:dio/dio.dart' show MultipartFile;
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:http_parser/http_parser.dart';
-import 'package:meta/meta.dart';
 import 'package:stream_chat/src/core/platform_detector/platform_detector.dart';
 import 'package:stream_chat/src/core/util/extension.dart';