+++ packages/stream_chat/lib/src/core/models/user.dart
@@ -41,6 +41,7 @@ class User extends Equatable {
     Map<String, Object?> extraData = const {},
     this.online = false,
     this.banned = false,
+    this.banExpires,
     this.teams = const [],
     this.language,
   })  : createdAt = createdAt ?? DateTime.now(),
@@ -67,6 +68,8 @@ class User extends Equatable {
     'last_active',
     'online',
     'banned',
+    'ban_expires',
+    'dashboard_ban_channel_cid',
     'teams',
     'language',
   ];
@@ -129,13 +132,17 @@ class User extends Equatable {
   )
   final bool banned;
 
-  /// Map of custom user extraData.
-  @JsonKey(includeIfNull: false)
-  final Map<String, Object?> extraData;
+  /// The date at which the ban will expire.
+  @JsonKey(includeIfNull: false, toJson: Serializer.readOnly)
+  final DateTime? banExpires;
 
   /// The language this user prefers.
   @JsonKey(includeIfNull: false)
   final String? language;
+
+  /// Map of custom user extraData.
+  @JsonKey(includeIfNull: false)
+  final Map<String, Object?> extraData;
 
   /// List of users to list of userIds.
   static List<String>? toIds(List<User>? users) =>
@@ -158,6 +165,7 @@ class User extends Equatable {
     bool? online,
     Map<String, Object?>? extraData,
     bool? banned,
+    DateTime? banExpires,
     List<String>? teams,
     String? language,
   }) =>
@@ -174,6 +182,7 @@ class User extends Equatable {
         online: online ?? this.online,
         extraData: extraData ?? this.extraData,
         banned: banned ?? this.banned,
+        banExpires: banExpires ?? this.banExpires,
         teams: teams ?? this.teams,
         language: language ?? this.language,
       );
@@ -186,6 +195,7 @@ class User extends Equatable {
         online,
         extraData,
         banned,
+        banExpires,
         teams,
         language,
       ];