+++ packages/stream_chat_flutter/lib/src/stream_chat_theme.dart
@@ -1,12 +1,5 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart' hide TextTheme;
-import 'package:stream_chat_flutter/src/channel_preview.dart';
-import 'package:stream_chat_flutter/src/gradient_avatar.dart';
-import 'package:stream_chat_flutter/src/message_input.dart';
-import 'package:stream_chat_flutter/src/reaction_icon.dart';
-import 'package:stream_chat_flutter/src/theme/themes.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
-import 'package:stream_chat_flutter_core/stream_chat_flutter_core.dart';
 
 /// Inherited widget providing the [StreamChatThemeData] to the widget tree
 class StreamChatTheme extends InheritedWidget {