+++ packages/stream_chat/CHANGELOG.md
@@ -1,13 +1,27 @@
 ## Upcoming
 
+🐞 Fixed
+
+- [[#857]](https://github.com/GetStream/stream-chat-flutter/issues/857) Channel now listens for member ban/unban and
+  updates the channel state with the latest data.
+- [[#748]](https://github.com/GetStream/stream-chat-flutter/issues/748) `Message.user` are now also included while saving users in persistence.
+
 🔄 Changed
 
-- `client.location` is now deprecated in favor of the new [edge server](https://getstream.io/blog/chat-edge-infrastructure) and will be removed in v4.0.0.
+- `client.location` is now deprecated in favor of the
+  new [edge server](https://getstream.io/blog/chat-edge-infrastructure) and will be removed in v4.0.0.
+- `channel.banUser`, `channel.unbanUser` is now deprecated in favor of the new `channel.banMember`
+  and `channel.unbanMember` and will be removed in v4.0.0.
 
 ✅ Added
 
 - Added `client.enrichUrl` endpoint for enriching URLs with metadata.
 - Fixed `unreadCount` after removing user from a channel.
+- Added `client.queryBannedUsers`, `channel.queryBannedUsers` endpoint for querying banned users.
+
+🐞 Fixed
+
+- [[#871]](https://github.com/GetStream/stream-chat-flutter/issues/871) Fixed thread message deletion.
 
 🐞 Fixed